+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Head
 
+- Added: `selector-no-qualifying-type` rule.
 - Fixed: `number-leading-zero` will not check `@import` at-rules.
 - Fixed: `selector-class-pattern` now ignores non-ouputting Less mixin definitions and called Less mixins.